# Greatly increase flotsam lifetimes and add drag