**Feature:** This PR implements a feature that's been talked about for a while now.

## Feature Details
Flotsams are outfits or commodities that have been jettisoned from ships or dropped by destroyed minable asteroids. These flotsams have been criticized as dying too quickly, and in the case of asteroids their flotsams have been criticized as moving too quickly to pick up.

This PR does two things:
1) Greatly increases the lifetime of flotsams. Commodity flotsams use to last anywhere from 6 to 11 seconds for a full box. Full commodity boxes now last between 2 and 3 minutes. Outfit flotsams on the other hand last 10 times longer than they use to. For example, ore flotsams use to last about 8 seconds. They now last about 80. (Flotsam lifetimes range from about 80 to 85 seconds on average depending on the type of minable, with Yottrite being the only suggested at about 65 seconds on average.) And in terms of more expensive outfits, a Triple Plasma Core that was jettisoned used to last 1.35 seconds. It now lasts about 13.5 seconds.
2) Added drag to flotsams. Flotsams now lose .1% of their velocity every frame. This means that after about 11 seconds, flotsams will have dropped to half their original velocity.

Questions:
* Should the loss in velocity be decreased? Longer lived flotsams like ores and commodities will always come to a stop with the current value.
* Should the loss in velocity change depending on the flotsam? For example, should all flotsams die after reaching say 25% velocity, meaning flotsams that are worth more slow down quicker?

## UI Screenshots
N/A

## Usage Examples
N/A

## Testing Done
Went mining for ores. Flotsams from minables last much longer than they use to, and not picking them up allows them to slowly come to a stop.
Jettisoned some commodities and observed that they lasted even longer than the minables did.
Jettisoned a Triple Plasma Core and observed that it lasted much shorter than the commodities or ores.

## Performance Impact
None.