@MasterOfGrey 2021-02-04
To answer the questions and thereby carry over my discord comments for the record, I will apply two answers to each. The ideal case that I think would be perfect (and do not have any real expectation of seeing), and my opinion on the practical case I think would have low impact on the proposed changes:
- Should the loss in velocity be decreased? Longer lived flotsams like ores and commodities will always come to a stop with the current value.
  - Ideally: I think that there should be a lower limit where drag stops applying, and that at all times there should be a vector applied to simulate the effect of it moving through the gravity well (either with an actual constant vector applied towards the sun or with a fudged vector to make it circle). I think that particularly for longer-lived ones like commodities this would look really good, and also you'd be able to get moving debris fields that all carry some echo of what the ship that was destroyed for them to exist was doing at the time. (Would look way cool in like Wah Ki, Cardax, or Automata space).
  - Practically: I think those things coming to a stop eventually isn't really a problem, and the current implementation works perfectly well with the 'space is a fluid' behaviour of ships in ES.
- Should the loss in velocity change depending on the flotsam? For example, should all flotsams die after reaching say 25% velocity, meaning flotsams that are worth more slow down quicker?
  - Ideally: If it could use the mass of the flotsam as an input to determining drag that would be really cool and fit in especially well with the "ideally" for the previous question.
    - Please god no lets not tie velocity to _value_ though. I will cry.
  - Practically: It's not at all necessary, it'd really just be a nice-to-have, but I think just using a different fixed drag value for each category of commodities/minables/outfits would at least allow a sharp-eyed player to prioritise what to chase if they're picking up from a destroyed ship.