@Zitchas 2021-02-04
Longer lived flotsam? Awesome! 

Having generic mineables last almost forever makes sense to me. Asteroids have been around for millions of years. Doesn't make sense that something blasted off them would suddenly melt away into space in mere seconds or minutes.

On the other hand, having things like, well, reactors blow up or vanish or whatnot after a short while also makes sense. I guess it'll come down to playtesting to see how it feels. I don't think raw numbers really convey the right feeling. 

That being said, having everything last longer both makes sense, sounds good, and is something I've wanted for ages now, so, thanks!