@Amazinite 2021-02-04
Well #5556 is one of the things I want to work on for the next major release, so if these values are too high then we should be able to more easily tweak them later instead of needing to always change the source code and recompile.